feat: implement user authentication improvements, including password reset functionality and accordion UI components

+++ frontend/src/routes/(auth)/login/+page.server.ts
@@ -6,42 +6,64 @@ import prisma from "$lib/server/prisma.server";
 
 export let actions = {
     login: async ({ cookies, request }) => {
-        // const formData = await request.formData()
-        // const username = formData.get("username")?.toString()
-        // const password = formData.get("password")?.toString()
+        let formData = await request.formData();
+        let username = formData.get("username")?.toString();
+        let password = formData.get("password")?.toString();
 
-        // if (!username || !password) {
-        //     return fail(400, { error: "Missing username or password" })
-        // }
+        if (!username || !password) {
+            return {
+                form: {
+                    errors: {
+                        username: "Invalid username",
+                        password: "Invalid password"
+                    }
+                }
+            }
+        }
 
-        // const user = await prisma.users.findUnique({
-        //     where: { username }
-        // })
+        // get user by username
+        let user = await prisma.users.findUnique({
+            where: { username }
+        });
+        // // ! bypassing password check for now
+        if (!user || !user.hashed_password) {
+            return {
+                form: {
+                    errors: {
+                        username: "Invalid username",
+                        password: "Invalid password"
+                    }
+                }
+            }
+        }
 
-        // if (!user || !user.hashed_password) {
-        //     return fail(401, { error: "Invalid username or password" })
-        // }
-
-        // const encoder = new TextEncoder()
-        // const raw = encoder.encode(password)
-        // const hashedPassword = await crypto.subtle.digest("SHA-256", raw)
-
-        // if (
-        //     !Buffer.compare(
-        //         Buffer.from(user.hashed_password),
-        //         Buffer.from(hashedPassword)
-        //     ) === 0
-        // ) {
-        //     return fail(401, { error: "Invalid credentials" })
-        // }
+        const encoder = new TextEncoder();
+        const raw = encoder.encode(password);
+        const hashedPassword = await crypto.subtle.digest("SHA-256", raw);
+        const isMatch = Buffer.compare(
+            Buffer.from(user.hashed_password, 'base64'),
+            Buffer.from(hashedPassword)
+        ) === 0
+        // console.log(isMatch, Buffer.from(user.hashed_password).toString(), '/n', Buffer.from(hashedPassword).toString())
+        // // ! bypassing password check for now
+        if (!isMatch) {
+            return {
+                form: {
+                    errors: {
+                        username: "Invalid username",
+                        password: "Invalid password not match"
+                    }
+                }
+            }
+        }
 
         await prisma.session.deleteMany({
             where: {
-                user_id: 1 //! Replace with the actual user ID from the login
+                user_id: user?.id
             }
         })
 
-        const session = await createSession(1)
+        const session = await createSession(user?.id ?? 1);
 
         cookies.set("session", session.token, {
             path: "/",
@@ -58,7 +80,7 @@ export let actions = {
         return {
             form: {
                 errors: {
-                    email: "Invalid email address",
+                    username: "Invalid username address",
                     password: "Password must be at least 8 characters long"
                 }
             }
